[BUGFIX] Fix script URLs for traditional modules
The introduction of CSRF protection for backend modules
introduced a regression for traditional modules.

Previously the script was calculated correctly if
not provided. Now API calls without providing the script
are assumed to be mod.php dispatched.

Fix that by checking if the M GET argument was sent
and fall back to traditional URL calculation.

Resolves: #62569
Releases: 6.2, master
Change-Id: I26d16acb513d678f117e0b645fe3eefebf909e3d
Reviewed-on: http://review.typo3.org/33626
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>