[!!!][TASK] Migrate Workspaces Toolbar JS to jQuery
In the main effort to replace Prototype.js with
simple RequireJS modules and jQuery, the Toolbar
Menu in the top of the workspaces extension
was replaced by a RequireJS alternative.

Additionally, this patch replaces the ExtDirect
call with a simple AJAX call by jQuery and ajax.php
to become ExtDirect-free for the module.

Additionally, the backend is now re-colored
again when inside a workspace, and the module
menu keeps its current active module when refreshing
dynamically via AJAX.

Resolves: #62800
Releases: master
Change-Id: Ie0bbe2c12dff6e1d3047cad5a175abc6f12393e4
Reviewed-on: http://review.typo3.org/33931
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>