[BUGFIX] RTE: Multi-line lists in default PageTS configs cause problems
Problem: For the sake of clarity, some of the RTE default PageTS
configurations specify lists as multi-line values. If any of these is
ever json-encoded, as some are, for use in JavaScript, some entries in
the list will start with a line feed.
Solution: Use addToList syntax for each additional line.

Releases: master, 6.2
Resolves: #62805
Change-Id: I6cb5eb6556adf2c61bd2610bb6653d767b5972c8
Reviewed-on: http://review.typo3.org/33935
Reviewed-by: Stanislas Rolland <typo3@sjbr.ca>
Tested-by: Stanislas Rolland <typo3@sjbr.ca>