[BUGFIX] Wrong generation of metadata in FAL
This patch fixes the generation of metadata if the metadata
of a file is missing in sys_file_metadata. Without this patch
the metadata is generated without taking care of images and their
width and height. As a result, images are generated with
width=0 height=0. See bugreport for details.

The change also introduces the basic structure to fetch
general info of files via the SplFileInfo objects API.

Resolves: #62829
Releases: master, 6.2
Change-Id: Ie99fc74f3439cef5dbd3702e57596b37390da679
Reviewed-on: http://review.typo3.org/36454
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>