[FOLLOWUP][BUGFIX] Use bootstrap JS as include instead of AMD
Bootstrap JS is included as AMD module,
thus breaking things like Extension Manager
which was coded in an uncool way, namely
not using API methods due to crappy Fluid
ViewHelpers for the backend.

For the time being and to keep everybody
happy, bootstrapJS is included as simple JS
file, but wrapped within TYPO3.jQuery to
use the previously loaded functionality.

Releases: master
Resolves: #62884
Change-Id: I71be335b39c1a2d1151f5e550bacb5643ee621d4
Reviewed-on: http://review.typo3.org/34011
Reviewed-by: Frank Nägler <typo3@naegler.net>
Tested-by: Frank Nägler <typo3@naegler.net>
Reviewed-by: Andreas Fernandez <andreas.fernandez@aspedia.de>
Tested-by: Andreas Fernandez <andreas.fernandez@aspedia.de>
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>