[!!!][TASK] Gifbuilder: Always use full filename of original file as prefix
As of now, images created with GIFBUILDER were created by a
hash in typo3temp/GB/. With the TypoScript option
config.meaningfulTempFilePrefix it was possible to prepend the
first X characters of the original filename in the filename.
This was due to some FS limitations on Windows.

With FAL the option was not used anymore, but Gifbuilder
still used this option.

In order to streamline the generation of temporary
files, images created by the Gifbuilder now contain
the full filename of the original file as well.

The TypoScript option "config.meaningfulTempFilePrefix"
has no effect anymore now.

Resolves: #62886
Releases: master
Change-Id: I746aa5d8c2e483212c28419dadb68a5cd5d1282d
Reviewed-on: http://review.typo3.org/34015
Reviewed-by: Eric Chavaillaz <eric@hemmer.ch>
Tested-by: Eric Chavaillaz <eric@hemmer.ch>
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>