[FEATURE] Let the reports in the install module check for new versions
This patch modifies the reports of the install module. Now it behaves
similar to the install tool and issues an error or a warning status
when there is a new version available.

Currently this function only works for the current minor version
(e.g. 6.2.X). It will not notify about updates to e.g. 7.X.X. Besides
it does not take care about development versions at all.

Version information is downloaded from get.typo3.org. If this is not
possible, the old functionality is still working (which is printing out
the version information only with Status::NOTIFY.

With the version information it is checked if the current version is
up-to-date. If it is, the version is printed out with Status::OK. If not,
it checks if there is a security issue in the current version. If yes, the
status get severity Status::ERROR, otherwise Status::WARNING is used.

Resolves: #62444
Related: #62649
Releases: master
Change-Id: Ifdb62e3a4c34f700a73087271d7bff7daf6b8838
Reviewed-on: http://review.typo3.org/33745
Reviewed-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Tested-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>