[BUGFIX] Correctly parse timestamps in format.date VH
If the timestamp is retrieved from child nodes, the date format
ViewHelper has to ensure the timestamp is properly recognized.
As child nodes are always returning a string, the former check with
is_integer() would never return TRUE.
Therefore adapt this check to use the MathUtility method as this
guarantees a proper check.

Resolves: #62932
Releases: master, 6.2
Change-Id: I6968991bf97f99f82fc942d7d17678c9b473e796
Reviewed-on: http://review.typo3.org/34323
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>