[BUGFIX] Make sys_file_metadata publishable
Currently sys_file_metadata records that are edited in a workspace
cannot be published because the workspace module does not know
about ignored root level restrictions and these records are on
root level by definition.

Include the root page when searching for workspace records
and show elements for which root level restriction is ignored.

Resolves: #60343
Releases: 6.2, master
Change-Id: I43a273661addff4a4e337bb464eb0cc4fd1fb823
Reviewed-on: http://review.typo3.org/34188
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Alex Kellner <alexander.kellner@einpraegsam.net>
Tested-by: Alex Kellner <alexander.kellner@einpraegsam.net>