[BUGFIX] Make sys_file_metadata publishable
Currently sys_file_metadata records that are edited in a workspace
cannot be published because the workspace module does not know
about ignored root level restrictions and these records are on
root level by definition.

Include the root page when searching for workspace records
and show elements for which root level restriction is ignored.

Resolves: #60343
Releases: 6.2, master
Change-Id: Ib82f5c7f7a67824c5d7d4e7cdaa01c89e0912731
Reviewed-on: http://review.typo3.org/34156
Tested-by: Gerhard Rupp <gerhard.rupp@respon.se>
Reviewed-by: Alex Kellner <alexander.kellner@einpraegsam.net>
Tested-by: Alex Kellner <alexander.kellner@einpraegsam.net>
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Reviewed-by: Helmut Hummel <helmut.hummel@typo3.org>
Tested-by: Helmut Hummel <helmut.hummel@typo3.org>