[BUGFIX] CSS3 function calc must retain whitespace
Whitespaces within CSS3 function 'calc' must not be stripped
because otherwise browsers won't recognize it anymore.

Although the CSS3 specification requires the whitespaces around
the additive expressions (+/-) only, it might be better to not
remove any whitespace within a calc function in terms of
simplicity.. the minifying is complex enough as it already is.

Resolves: #62463
Releases: master, 6.2
Change-Id: Ied0c02e132aafa97ce9fb6b0e9930898cb17efc1
Reviewed-on: http://review.typo3.org/34180
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>