[BUGFIX] Fix JS errors in page module
Due to the introduction of the fontawesome
icons drag&drop was not possible anymore
with ExtJS throwing errors in the page module
when starting to drag.

The intermediate fix is to make sure that
no errors are thrown in page module.

The correct fix is to define proper CSS
classes to address the elements,
but not via icon classes.

Releases: master
Resolves: #62969
Change-Id: I67b86f2b38054093b8755d6f9f65d868b3ccff76
Reviewed-on: http://review.typo3.org/34194
Reviewed-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Tested-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>