[CLEANUP] Deprecate unused/non-unified DocumentTemplate code
Several functions within DocumentTemplate are
not encouraged to be used anymore.

The hard-coded background image setting
via $TBE_STYLES['background'] is removed
and its usage is deprecated.

Additionally, the font-wrapping methods
are deprecated. All places in the core using
these methods are removed. The according
CSS was removed from the core (can still
be used via 3rd party extensions).

collapseableSection, which was used solely
by the reports module for ages in a buggy
way, is also deprecated in favor of Bootstrap
collapseables and localstorage.

Change-Id: If1abdeb1d9491efb385b41eb4d1faf4e1dbc91e0
Resolves: #62988
Releases: master
Reviewed-on: http://review.typo3.org/34209
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>