[CLEANUP] Remove supersuperold JS file
There is a file in typo3/js/jsfunc.updateform.js
which was used before all fe_admin was
removed from core, but missed from removal
with fe_admin completely.

Same goes for the call in ContentObjectRenderer
which is not used in the core for several versions.

As this can not work, and is superfluous, the file
and the method (note: @access private) can
be safely removed.

Releases: master
Resolves: #62996
Change-Id: I8f30d6d6ca317c616e05aa6ee88efb73adc15033
Reviewed-on: http://review.typo3.org/34219
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>