[TASK] Remove leftover modulemenu JS
Prior to the module menu restructuring
based on jQuery, the module menu
was a ExtJS viewport to be fit in the
backend scaffolding dynamically built.

When adding the new code, the old
wasn't completely removed. A ExtJS
panel named modulepanel was still
included but is not in use anymore.

Resolves: #62998
Releases: master
Change-Id: I367bbce1a7dc35b62e0ecaa8c5f16765210e0974
Reviewed-on: http://review.typo3.org/34221
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>