[BUGFIX] Allow frontend requests from different document root
With a specially prepared index.php and using a page renderer hook
it is in general possible to deliver frontend requests
from a dedicated web folder.

However there are a few (hard coded) issues in TYPO3 which makes
this impossible.

Change GeneralUtility::getIndpEnv('TYPO3_SITE_URL') to check for
a TYPO3_PATH_WEB constant defined in a specially prepared index.php
in a "Web/" folder and provide the fluid stand alone view with
an absolute path (which is meant like that in the API anyways).

This change has no impact on a TYPO3 installation with a traditional
directory structure, but enables users to have a more secure setup
in an appropriate environment.

Resolves: #63025
Releases: 6.2, master
Change-Id: I571ceb6224c8030d1f5f99a45903cda3a8171313
Reviewed-on: http://review.typo3.org/34288
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>