[BUGFIX] Distributions: Import files before database
The distribution import does sometimes not import the Assets
(Resources/Files/) of the distribution. This is because the
t3d-file is imported before the files are imported.
That means that the t3d import will create FAL file relations
to files that do not exists. This will end up in an FAL exception.
Files should be imported before the t3d-file is imported.

Resolves: #63121
Releases: master,6.2
Change-Id: Ibb3695b0b92ed2a257380c8684e4451851a2248f
Reviewed-on: http://review.typo3.org/36466
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>