[BUGFIX] EXT:lang: Hide useless UI if no language is selected
In EXT:lang, the update button is displayed if no language is selected.
Clicking on that button will display several FlashMessages and PHP
throws an exception due to missing arguments.

This patch takes care of it and simply hides the button if no
language is selected.
Furthermore it hides the righthand side of the table too.

Resolves: #63124
Releases: master
Change-Id: Ica69a257ff5f36eda789051b194ef33c47e699fc
Reviewed-on: http://review.typo3.org/34436
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Felix Kopp <felix-source@phorax.com>
Reviewed-by: Tymoteusz Motylewski <t.motylewski@gmail.com>
Tested-by: Tymoteusz Motylewski <t.motylewski@gmail.com>