[TASK] Migrate ViewPageIcon to Sprite Icon
There are still some small places where
sprite icons are not used yet.

One is the link to the view page icon.
The call within document template
is only used twice in the core:
1. DocumentTemplate->getHeader()
2. ElementInformationController

The patch replaces the hard-coded
image tag with a sprite icon.

As this function is hardly used,
you need to know where to test:
- Go to the Info module
=> PageTSconfig
=> ViewTSconfig fieldContent
=> Choose a page in the pagetree which
is not the root page.
On the bottom you see the TSconfig
for each page, with the search
icon and the new icon once
the patch is applied.

Releases: master
Resolves: #63138
Change-Id: I8d2eaf8fe9a6cf8a6664fcaa33223ac72d637795
Reviewed-on: http://review.typo3.org/34464
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>