[TASK] Move wizards to Web=>Functions
The current situation of the Functions module
is like this:
Wizards like "Sort pages" and
"Bulk-create pages" are inside
Web=>Functions=>Wizards.
The Functions module is not used
at all for anything else right now.

For a cleaner interface and code
structure, both wizards are moved
to the Web=>Functions level directly.

This avoids having a separate dropdown
in the content area, but using the already
known (and default) "functions" MOD MENU.

In a further step, the "func_wizard" extension
should be moved to compatibility6 with a
proper note on how to migrate old "wizard"
extensions to the normal top-level functions
module.

Resolves: #63310
Releases: master
Change-Id: I9845cafe87de3f2f93d7938bef28284b80a04022
Reviewed-on: http://review.typo3.org/34581
Reviewed-by: Stefan Froemken <froemken@gmail.com>
Tested-by: Stefan Froemken <froemken@gmail.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>