[BUGFIX] Properly check permissions with read only file mounts
Previously the permission check returned false if the
file or folder was within a read only file mount, but
also within a writable file mount.

Properly check this now and centralize the check in
isWithinFileMountBoundaries, which now has an additional
argument whether write access to file mounts should be checked or not.

Resolves: #63364
Related: #49391
Releases: master, 6.2
Change-Id: If90958b7d6e4d5aad1fbd172af06462ca2b9764f
Reviewed-on: http://review.typo3.org/34964
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>
Reviewed-by: Alexander Stehlik <alexander.stehlik@gmail.com>
Tested-by: Alexander Stehlik <alexander.stehlik@gmail.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>