[TASK] Remove jquery tools from contrib
jQuery tools was added for
the EXT:documentation module,
however it is not used at all
in the core.

Additionally, we don't recommend
using it in the backend either.

The file contents do not state anything
about the version or license which
is a big problem for anyone to use
it.

The obsolete and unused file is
therefore removed.

Resolves: #63436
Releases: master
Change-Id: I7e8caa86bcf9859e5b545bf018e2b6e5c9e80e87
Reviewed-on: http://review.typo3.org/34787
Reviewed-by: Frank Nägler <typo3@naegler.net>
Tested-by: Frank Nägler <typo3@naegler.net>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>