[BUGFIX] Redirect to distributions after first login does not work
When redirecting to the distributions modulefunction
after the first login, it does not work, as the backend only
redirects to the module itself, but does not add parameters
to the module loading instance.

The patch makes sure that the startModule also
adds the second parameter.

For review, have a look at
BackendController->setStartupModule
and EXT:install/Action/Step/DefaultConfiguration
(look for startModuleOnFirstLogin)
to understand the logic.

Releases: master
Resolves: #63521
Change-Id: I4280717556b097ddd242e45fe42803b74814eb47
Reviewed-on: http://review.typo3.org/34944
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>