[TASK] Add configuration file for DynamicReturnTypePlugin
Adding configuration file for DynamicReturnTypePlugin for PHPStorm.
The plug-in makes it possible to use code completion for dynamic
return types. plug-in allow a dynamic return type from method calls.
It can use either the instance type of the passed parameter or a
string look-up.

Since the ::class feature is used this is a good addition for all
developers.

PHPStorm Plugin: http://plugins.jetbrains.com/plugin/7251

Resolves: #63553
Releases: master
Change-Id: I171de5400a213a375a614844e6f2671b74bcc938
Reviewed-on: http://review.typo3.org/35006
Reviewed-by: Jigal van Hemert <jigal.van.hemert@typo3.org>
Tested-by: Jigal van Hemert <jigal.van.hemert@typo3.org>