[BUGFIX] Use die() instead of exception for PHP version check
The installer checks for the PHP version and throws and exception.
This is bad for people not having enabled display_errors in their
PHP config and leads to an empty white page.

Change this to simply die() as this will ensure the user can
see the cause of the issue.

Resolves: #63570
Releases: master
Change-Id: I84f5f1e089753beffe17360a636d2d5038ab24a6
Reviewed-on: http://review.typo3.org/35058
Reviewed-by: Alexander Opitz <opitz.alexander@googlemail.com>
Tested-by: Alexander Opitz <opitz.alexander@googlemail.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Frank Nägler <typo3@naegler.net>
Tested-by: Frank Nägler <typo3@naegler.net>