[TASK] Use autoload map instead of aliases in sysext form
Using the alias map is misleading as it is only used
to force the class loader to cache class names with
lowercased identifier to be able to load the class
with "wrong" casing.

Use the autoload map for that instead and clarify
in that file that it must not be removed for now.

Resolves: #63999
Related: #63576 #63577
Releases: master
Change-Id: I8f683ff1d68480a3530a8d2fea62208e66ca947d
Reviewed-on: http://review.typo3.org/35599
Reviewed-by: Helmut Hummel <helmut.hummel@typo3.org>
Tested-by: Helmut Hummel <helmut.hummel@typo3.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>