[TASK] Replace BrowseLinks.js with jQuery/RequireJS Module
In the main approach to drop prototype/scriptaculous
usage, the BrowseLinks functionality
is rewritten to be used as a RequireJS module.

As BrowseLinks is still a lot of really old code
this needs to be refactored more at a later stage.

To test: FAL / IRRE usage, wizard
functionality, RTE functionality.

Resolves: #63690
Releases: master
Change-Id: I58514fa83a627fe34d13dd77ece0c44f495ee45d
Reviewed-on: http://review.typo3.org/35173
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Reviewed-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Tested-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>