[BUGFIX] Better performance for bulk inserts without static_write
In DataHandler->fillInFieldArray for every record to insert or update
current and new data are merged to be handed over to the static_write
functionality. If this functionality is disabled (which is the
default) the merged data is discarded again.

Checks if the static_write functionality is enabled and merges data
only if it is enabled. This drastically lowers memory usage for bulk
operations.

In master, the static_write functionality was removed completely,
for 6.2 the performance of this code section is improved.

Resolves: #63668
Releases: 6.2
Change-Id: I625a771a98a5f79bab3a8d22464895fcbdcf4068
Reviewed-on: http://review.typo3.org/35370
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>