[!!!][TASK] Remove words and words_strings from ReferenceIndex
In ReferenceIndex->generateRefIndexData the strings from all input and
text fields were collected in $this->word_strings. But the array was
never accessed.

While both properties were removed in master, the 6.2 backport keeps
the properties, but does not fill word_strings anymore, so the patch
is less breaking but the performance improvement is still given.

Resolves: #63780
Releases: master, 6.2
Change-Id: I56d6a3e615cab44cbecff5a7d4dec42a1da29fc6
Reviewed-on: http://review.typo3.org/35349
Reviewed-by: Stephan Großberndt <stephan@grossberndt.de>
Tested-by: Stephan Großberndt <stephan@grossberndt.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>