[!!!][TASK] Remove words and words_strings from ReferenceIndex
In ReferenceIndex->generateRefIndexData the strings from all input and
text fields were collected in $this->word_strings. But the array was
never accessed. Remove ReferenceIndex->words (unused) and
ReferenceIndex->word_strings and the collection of word_strings. This
greatly reduces memory usage when bulk inserting or editing records as
well as updating the reference index.

Resolves: #63780
Releases: master, 6.2
Change-Id: I56d6a3e615cab44cbecff5a7d4dec42a1da29fc6
Reviewed-on: http://review.typo3.org/35310
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>