[BUGFIX] Fixes position of suggest wizzard autocomplete result in IE
A 5 years old bug in script.aculo.us leads to the problem, that
autocomplete results of the suggest wizzard are displayed at a
wrong position for Internet Explorer browsers. Depending on the
screen resolution, the autocomplete results may even be out of the
visual area of the users screen.

Since script.aculo.us seems to be unmaintained, the patch is directly
applied to the script.aculo.us source files, so we at least can
provide a bugfix for TYPO3 users using Internet Explorer.

This patch also fixes an IE fix in script.aculo.us. Instead of using
javascript:false we now use about:blank as source for the IE fix
iframe, which prevents the text “false” being displayed when
autocompletion results appear in IE.

Resolves: #63604
Releases: master, 6.2
Change-Id: Ib0421a308bb0652b9aed8be73ab6c78900ebe13d
Reviewed-on: http://review.typo3.org/35786
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>