[BUGFIX] DataHandler: Cache TCA field eval array
Introduce a runtime cache for the eval values of TCA input and text
fields. This improves performance for bulk inserts since
DataHandler->checkValue_input and checkValue_text are called for each
according field of each record and the available evaluations do not
change during a call. If no eval is set, do not execute
checkValue_*_Eval functions at all.

Additionally align usages of getMemoryCache() to runtimeCache and
keys of nestedElementCalls and elementsToBeDeleted in cache_runtime.

Resolves: #63670
Releases: master, 6.2
Change-Id: I3d2122a61e3194402743fa0b364d6bd80d4e72dc
Reviewed-on: http://review.typo3.org/35186
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>
Reviewed-by: Jigal van Hemert <jigal.van.hemert@typo3.org>
Tested-by: Jigal van Hemert <jigal.van.hemert@typo3.org>