[FEATURE] Introduce unified Storage API for JS
The core lacks a unified API to load and update
the Backend users UC, and to address the browsers'
localStorage.

The new TYPO3.Storage object allows to fetch
and manipulate data in a unified way.

Followup patches will replace previous hacks
and ExtJS calls with the new API.

Releases: master
Resolves: #64031
Change-Id: I53907f0beb93f300f8d8ca7a17721cab1376b9c7
Reviewed-on: http://review.typo3.org/35627
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>