[BUGFIX] Editing a record in Element Browser throws JS error
When opening the element browser (e.g. via the field "reference
records" as CType in tt_content), it is actually possible to edit
all title fields / headlines of a record. This is done after
sorting a field. However, some JS variables are not set in the
element browser.

Resolves: #19830
Releases: master
Change-Id: Ic2058a825dea1b2ef18c049bfff92eb94de630df
Reviewed-on: http://review.typo3.org/35677
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>