[BUGFIX] Sections not properly handled by record collector
When using the Unit-Test framework frontend record collector the
resulting response is not correct.

Instead of adding records only to the specific section (configured via
TypoScript) the records pile up in $this->records and get added to
every section. The reason this only true for $this->records of the
Collector is, that for the other class variables (like $this->structure)
there are appropriate array keys used anyways.

Solution is to reset the internal member variables used for collecting
record information each time after attaching a section.

Resolves: #64066
Releases: master
Change-Id: I7a1ff945e0d267f1e0c6608637479ea0cc28fe70
Reviewed-on: http://review.typo3.org/35692
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Mateusz Wojtuła <matw88@gmail.com>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>