[TASK] HMENU: Extract menu item generation into separate methods
In order to ensure valid options and consistency, the 500+ lines
of code for generating special=XY menus in HMENU
can be safely moved into separate methods within the same
class. Additionally, when subclassing HMENU, the functions
can be overridden easier.

The makeMenu() function gets smaller and the new methods are
a good target for creating unit tests.

Releases: master
Resolves: #57366
Change-Id: I1c5be14e4249fe60f7c14f808b762b2af3155680
Reviewed-on: http://review.typo3.org/28900
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Reviewed-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Tested-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>