[TASK] Cleanup EXT:viewpage JavaScript
The Extension viewpage (Web => View)
is still using a mixture of ExtJS and jQuery.

In order to use the logic from the new
Storage API (which replaces the ExtJS Part
for this module) the whole JavaScript
was rewritten in RequireJS.

The shipped jquery-ui JavaScript file
is not necessary anymore, as it is
loaded from the contrib/ directory
as a AMD module (new version).

Resolves: #64103
Releases: master
Change-Id: Ic0cc1dd0eb56b0440a5d8c78c40ac2120aa0f215
Reviewed-on: http://review.typo3.org/35767
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>