[BUGFIX] Avoid makeInstance calls starting with a backslash
When using singletons and calling makeInstance
with a beginning backslash in the class name
(see forge example), the singleton is not
correctly used.

Like inside the class loader, makeInstance and
the Extbase object manager should check for
a beginning backslash in the string of the
class name, and throw an exception, to give developers
a hint about broken code.

Change-Id: Ie7e0a6436fa8c5d983cd7e61b6ba96184440c657
Resolves: #60174
Releases: 6.2
Reviewed-on: http://review.typo3.org/35836
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>