[FEATURE] Allow appending of element browser mounts
A new User TSconfig option is introduced that allows the alternative
element browser page tree mounts to be appended instead of beeing
replaced:

options.pageTree.altElementBrowserMountPoints.append

Additionally the handling of temporary mount points within the RTE
and the wizard element browser is unified. If a temporary mount
point is active the altElementBrowserMountPoints will be ignored.

Resolves: #50780
Documentation: #63009
Releases: master
Change-Id: I2a0b616b398131684a2c7c71a25b4b51ef6f0c50
Reviewed-on: http://review.typo3.org/31898
Reviewed-by: Alexander Stehlik <alexander.stehlik@gmail.com>
Tested-by: Alexander Stehlik <alexander.stehlik@gmail.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>