[BUGFIX] Fix query in abbreviation plugin of rtehtmlarea
The comma seperated list of PIDs is not wrapped in quotes any more
to allow more than one page to be searched in.

Additionally invalid SQL queries with an empty IN parameter are
prevented by avoiding any query if the PID list is empty 
and the user is locked to DB mounts.

Resolves: #64169
Releases: master, 6.2
Change-Id: I3cf35a900ea84a5f7110e8b262f7aaccb27f3267
Reviewed-on: http://review.typo3.org/35864
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Reviewed-by: Stanislas Rolland <typo3@sjbr.ca>
Tested-by: Stanislas Rolland <typo3@sjbr.ca>