[BUGFIX] Prevent ResourceDoesNotExistException
An exception is thrown if the show info popup is opened by sys_file uid
for a file which does not exist anymore in filestorage.

This change marks the file as missing so no exception is thrown in
ElementInformationController but a error message is shown to the user
about the missing file.

Resolves: #64214
Releases: master, 6.2
Change-Id: I460013e21db149760a1b391d9cdbd68b6530d190
Reviewed-on: http://review.typo3.org/42594
Reviewed-by: Daniel Goerz <ervaude@gmail.com>
Tested-by: Daniel Goerz <ervaude@gmail.com>
Reviewed-by: Frans Saris <franssaris@gmail.com>
Tested-by: Frans Saris <franssaris@gmail.com>