[BUGFIX] Make TCEMAIN.translateToMessage work with lang files
The option TCEMAIN.translateToMessage that prepends
a certain text to a message was apparantly available
with the functionality of "LLL:" in the past.

TCEMAIN.translateToMessage = LLL:EXT:foobar/locallang_be.xml:translateTo

The patch re-introduces the functionality
again which was missing since at least 4.7.

See forge for further details.

Resolves: #41906
Releases: master, 6.2
Change-Id: I7f30ba7fc1d6577073e8a75840762c0912ca5d2c
Reviewed-on: http://review.typo3.org/35967
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>