[BUGFIX] Set correct workspace ID when updating referenceIndex for files
When publishing a workspace the referenceIndex entries of the affected
sys_file_reference rows were not updated with the new workspace. That's
because the user is always in the affected workspace and therefore
the workspace is not set to live as expected.
Now updating referenceIndex for both workspaces when swapping.

Resolves: #64253
Releases: master, 6.2
Change-Id: Icfb4fcd1ae92da4ea881105b387c3d6db20ece95
Reviewed-on: http://review.typo3.org/36481
Tested-by: Michael Schams <typo3@2014.trash.schams.net>
Reviewed-by: Stephan Großberndt <stephan@grossberndt.de>
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>