[TASK] Move getFieldDefaultValue to FORM cObject
The method getFieldDefaultValue() inside
ContentObjectRenderer is part of the FORM
content object and is called there.
Thus, it is moved to the FormContentObject
class as a protected method. All calls to this
method are changed accordingly.

As the function is marked for private access,
it is moved without any further documentation
or deprecation.

Resolves: #64391
Releases: master
Change-Id: Ib99503973412b34e31eac976117b6d707f5b1577
Reviewed-on: http://review.typo3.org/36149
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>