[BUGFIX] Fix class cache corruption in chroot environment
This commit fixes an issue that occurs in the rare edge case when the
TYPO3 cli dispatcher is called in a chroot environment and the site
is delivered via a non-chrooted web server.

In this case, all classes in the class cache will be stored with
`PATH_typo3` as base path, except the `TYPO3\Flow` classes in the
"core" extension. These are based on the __DIR__ constant, which
according to [1] resolves symlinks that generate directory paths that
are invalid outside of the chroot environment.

[1] http://php.net/manual/en/language.constants.predefined.php

Change-Id: I34a15c7a9913afc11cdd1d179a14331b157cbd29
Releases: 6.2
Fixes: #64508
Reviewed-on: http://review.typo3.org/36301
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>