[BUGFIX] Change highlight color on folder change
Clicking on folder in filelist results in JS error.

When using the filelist main frame
for navigation through the folder structure
there is still some leftover from the old
code, which changes the highlighted
item when clicking on a folder.

The patch solves the issue by properly
addressing the navigation frame and
the JS method.

This part is actually a leftover from pre-4.2
code which was not changed by me in 4.2
with the introduction of the AJAX-based
folder tree with prototype.js.

Resolves: #64572
Releases: master, 6.2
Change-Id: Iff7c7aff07454e428fe78194532ba95df2bd25d0
Reviewed-on: http://review.typo3.org/36342
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Jan Helke <typo3@helke.de>
Tested-by: Jan Helke <typo3@helke.de>
Reviewed-by: Alexander Opitz <opitz.alexander@googlemail.com>
Tested-by: Alexander Opitz <opitz.alexander@googlemail.com>
Reviewed-by: Tymoteusz Motylewski <t.motylewski@gmail.com>
Tested-by: Tymoteusz Motylewski <t.motylewski@gmail.com>