[BUGFIX] Keep linktype on core update
Keep the linktype of the symlink pointing to the TYPO3 core. Currently
the CoreUpdateService always creates an absolute link even if the
link was relative before. This patch checks for the correct link type
during a core update. This could be important in some environments,
e.g. in a chrooted environemnt.

Change-Id: I47bbe0d54a88853adc4160937a21d641b756f20f
Resolves: #58000
Releases: master, 6.2
Reviewed-on: http://review.typo3.org/36345
Reviewed-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Tested-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Reviewed-by: Göran Bodenschatz <coding@46halbe.de>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>