[BUGFIX] Fix handling of TCA allowed flag in ElementBrowser
If the 'allowed' flag in the element browser contains a '*', which
means any type of file, no file is displayed. This is because a filter
is added that only files are displayed that have a '*' as file ext.

The filter should not be added if the allowed value is '*' or empty.

Change-Id: I63016050a80c4f9d9bc0a4cd7c0dc7523ad9f58f
Resolves: #54504
Documentation: #64609
Releases: master, 6.2
Reviewed-on: http://review.typo3.org/36439
Reviewed-by: Philipp Gampe <philipp.gampe@typo3.org>
Tested-by: Philipp Gampe <philipp.gampe@typo3.org>