[BUGFIX] Return InaccessibleFolder for non existing processed folder
When the processed folder of a folder isn't present the ResourceStorage
tries to create it. When storage is offline or write-only this will fail
and as a result you run into exceptions on several places in BE.

This change makes sure there is a InaccessibleFolder object returned
in these cases.

Resolves: #61688
Releases: master, 6.2
Change-Id: Ie32c09f8d322914234860903c18ca44ff5cbf5db
Reviewed-on: http://review.typo3.org/32836
Reviewed-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Tested-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Reviewed-by: Fabien Udriot <fabien.udriot@ecodev.ch>
Tested-by: Fabien Udriot <fabien.udriot@ecodev.ch>
Reviewed-by: Frans Saris <franssaris@gmail.com>
Tested-by: Frans Saris <franssaris@gmail.com>
Reviewed-by: Michael Oehlhof <typo3@oehlhof.de>
Tested-by: Michael Oehlhof <typo3@oehlhof.de>