[BUGFIX] Prevent folder deletion with referenced files
Currently a folder can be deleted without any check if included files
are still in use anywhere within the website. This can lead to lost
files and can be harmful for the user. The patch adds a test if files
within a folder, which should be deleted, are still referenced and
prevents deleting it if files where found. The user is notified with a
warning shown above the file list.

Resolves: #48893
Releases: master, 6.2
Change-Id: If1c79d86e1f6c7a32a6a3e94e639951f882302d7
Reviewed-on: http://review.typo3.org/36413
Reviewed-by: Ingo Schmitt <is@marketing-factory.de>
Tested-by: Ingo Schmitt <is@marketing-factory.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Armin Ruediger Vieweg <armin@v.ieweg.de>
Tested-by: Andrea Herzog-Kienast <a.herzog@kienastdv.de>
Reviewed-by: Thomas Deuling <tdeuling@gmail.com>
Tested-by: Thomas Deuling <tdeuling@gmail.com>
Reviewed-by: Jan Helke <typo3@helke.de>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>