[BUGFIX] Failing unit test in SqlSchemaMigrationServiceTest
Currently the test for handling multiple white spaces within the sql
field definitions fails on Windows systems because of a different key
for the table name within the expected and actual array. This is caused
by a strict lowercase converting in the code for Windows systems. As the
table name is irrelevant for the test itself, this patch simply converts
table name to lower case.

Releases: master, 6.2
Resolves: #64657
Change-Id: I8d083e988b9a7403f0ff514af1ebe54adf065383
Reviewed-on: http://review.typo3.org/36463
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>