[BUGFIX] Add only required classes and class aliases
The static class and alias map now only contains a class map
and an alias map for required extensions.

This means that the traditional class loader will kick in for other
activated extensions.

Optionally it is possible to provide additional class and alias maps
in a distribution composer.json, this will however only work
if the TYPO3_COMPOSER_AUTOLOAD environment variable is set.

The latter is a precaution to not break backwards compatibility.

Releases: master, 6.2
Resolves: #64376
Change-Id: I4fd2202abad345c29bcb2b05d3702bdeef9e4474
Reviewed-on: http://review.typo3.org/36460
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>