[BUGFIX] ExtractorRegistry can not handle Extractors with same priority
The ExtractorRegistry::getExtractors() overrides existing
instances with the same execution priority. There is no check
or notification about this. As instance, it could be that you
have a extractor for Local driver and an extractor for a remote
driver with the same execution priority, that is currently not possible.

The change set also keep in sync some part of the code with
the FileRenderRegistry which is very close to the Extractor Registry.

Change-Id: I42fdd3916410536a0b097d96cf833129f5359d72
Releases: master, 6.2
Resolves: #56727
Reviewed-on: http://review.typo3.org/36555
Reviewed-by: Philipp Gampe <philipp.gampe@typo3.org>
Tested-by: Philipp Gampe <philipp.gampe@typo3.org>