[BUGFIX] Processing folder has now nested subfolders
In case of big installations the amount of processed files in
one folder is too much and causes performance issues and other
problems.
To prevent this kind of issues, nested subfolders are now
created in the processing folder in order to split the processed
files into more than one folder.

Resolves: #56557
Releases: master, 7.6
Change-Id: Id9a4fc3b4bb4b28ca26ff96221097171c3835eb7
Reviewed-on: https://review.typo3.org/47854
Reviewed-by: Philipp Gampe <philipp.gampe@typo3.org>
Tested-by: Philipp Gampe <philipp.gampe@typo3.org>