[BUGFIX] Scheduler: Show exception when invalid CommandController present
When a CommandController in an extbase extension has been registered, 
but the class does not contain the string "CommandController"
an exception will be thrown. Before this patch, this exception
will never show up.

This patch adds the output of exception's message in Scheduler module,
which gives the developer a hint why the "add new task" form is not
displayed properly.

Also it extends the exception with description why the class name is invalid.

Resolves: #64836
Releases: master, 6.2
Change-Id: I309ed02c7fd4597b99388518abf842e600b27a9a
Reviewed-on: http://review.typo3.org/36637
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Reviewed-by: Alexander Opitz <opitz.alexander@googlemail.com>
Tested-by: Alexander Opitz <opitz.alexander@googlemail.com>
Tested-by: Markus Klein <klein.t3@reelworx.at>