[BUGFIX] Make composer class loading case insensitive
To fix an edge case where wrong casing of a class name
in third party code did not produce a fatal error in TYPO3 6.2,
we change the composer class loader which is shipped by default
to allow case insensitive class loading of core classes.

This behavior can be switched on and off for fully composer based projects.
It is off by default (which will be the case in master)
and on in TYPO3 6.2 for full backwards compatibility.

Releases: 6.2
Resolves: #64873
Change-Id: I9b4de69fdd2b56616cc7ca249395a22f242b0c33
Reviewed-on: http://review.typo3.org/36670
Reviewed-by: Helmut Hummel <helmut.hummel@typo3.org>
Tested-by: Helmut Hummel <helmut.hummel@typo3.org>
Reviewed-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Tested-by: Mathias Schreiber <mathias.schreiber@wmdb.de>