[BUGFIX] Indexed Search pibased plugin can load again translations
This patch fixes a regression introduced in bf63a1c,
which changed the variable $scriptRelPath with a wrong
file location. The old file name is reverted so that
AbstractPlugin will look for the locallang.xlf file at
the correct location.

The patch also reformats the properties comments of the class.

Reference: https://review.typo3.org/#/c/32708/

Resolves: #64879
Releases: master
Change-Id: I58a07f11e382bde2ed84835ca0820ff5e00e1998
Reviewed-on: http://review.typo3.org/36677
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Daniel Maier <dani-maier@gmx.de>
Tested-by: Daniel Maier <dani-maier@gmx.de>
Reviewed-by: Göran Bodenschatz <coding@46halbe.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>