[BUGFIX] Implement check against File type restriction in Extractors
The Metadata Extractor interface provides the possibility to limit
the scope of the extraction to only certain file types, e.g. text, image,
video, audio, application. This can be defined in every Extactor
in method ExtractorInterface::getFileTypeRestrictions()

However, this check is not implemented in the Indexer and defining
any value as File type restriction will have no effect currently.
This patch fixes the situation.

Change-Id: Iffc01f98eec89de818acaa8ceb759a2b781759fa
Resolves: #64883
Releases: master, 6.2
Reviewed-on: https://review.typo3.org/36674
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Reviewed-by: Daniel Goerz <ervaude@gmail.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>